requirements.txt
pandas
llama-cpp-python